Move models into team
